Fixed decimals display for the prices fields and dataGridView names, the coresponding stored procedures were modified. Completed the report viewer, with the field Garantie still needing some work and the design of the report.

+++ ledger_class.cs
@@ -40,11 +40,11 @@ namespace AplicatieDisertatie
 
         public string observatii { get; set; }
 
-        public decimal pret_achitat { get; set; }
+        public int pret_achitat { get; set; }
 
-        public decimal pret_avans { get; set; }
+        public int pret_avans { get; set; }
 
-        public decimal pret_estimativ { get; set; }
+        public int pret_estimativ { get; set; }
 
         public string termen_rezolvare { get; set; }
